requirements.txt
Flask
Flask-SQLAlchemy
numpy
opencv-python
Pillow
scikit-learn
joblib
Werkzeug
matplotlib
seaborn
